+++ src/view/client/index.js
@@ -1,0 +1,3 @@
+import View from './ClientView'
+
+export default View