+++ src/view/server/index.js
@@ -1,0 +1,3 @@
+import View from './ServerView'
+
+export default View